appimage build fix

+++ _packaging/appimage/build.sh
@@ -10,4 +10,4 @@ chmod +x ./pkg2appimage
 ./pkg2appimage ./VidCutter.yml
 rm ./pkg2appimage
 
-mv out/VidCutter--${ARCH}.AppImage out/VidCutter-${VERSION}-x64.AppImage
+[ -f "out/VidCutter--${ARCH}.AppImage" ] && mv out/VidCutter--${ARCH}.AppImage out/VidCutter-${VERSION}-x64.AppImage
